# gravitee-apim-rest-api/gravitee-apim-rest-api-management-v2/gravitee-apim-rest-api-management-v2-rest/src/main/resources/openapi/openapi-apis.yaml
openapi: 3.0.3
info:
    title: Gravitee.io APIM - Management API - APIs
    description: |-
        This is the OpenAPI specification for our new version of APIM Management API.
    contact:
        email: team-apim@graviteesource.com
    license:
        name: Apache 2.0
        url: http://www.apache.org/licenses/LICENSE-2.0.html
    version: 2.0.0
security:
    - BasicAuth: []
    - CookieAuth: []
servers:
    - url: "{protocol}://{managementAPIHost}/management/v2"
      description: APIM Management API v2 - Default base URL
      variables:
          protocol:
              description: The protocol you want to use to communicate with the mAPI
              default: https
              enum:
                  - https
                  - http
          managementAPIHost:
              description: The domain of the server hosting your Management API
              default: localhost:8083
    - url: "{protocol}://{managementAPIHost}/management/v2/organizations/{orgId}"
      description: APIM Management API v2 - Base URL to target specific organizations
      variables:
          protocol:
              description: The protocol you want to use to communicate with the mAPI
              default: https
              enum:
                  - https
                  - http
          managementAPIHost:
              description: The domain of the server hosting your Management API
              default: localhost:8083
          orgId:
              description: The unique ID of your organization
              default: DEFAULT
...
paths:
...
    /environments/{envId}/apis/{apiId}/subscriptions/{subscriptionId}:
        parameters:
            - $ref: "#/components/parameters/envIdParam"
            - $ref: "#/components/parameters/apiIdParam"
            - $ref: "#/components/parameters/subscriptionIdParam"
        get:
            tags:
                - API Subscriptions
            parameters:
                - name: expands
                  in: query
                  description: Expansion of data to return in subscriptions.
                  schema:
                      type: array
                      items:
                          type: string
                          enum:
                              - application
                              - plan
                              - subscribedBy
            summary: Get one API's subscription
            description: |-
                Get the API's subscription by its identifier.

                User must have the API_SUBSCRIPTION[READ] permission.
            operationId: getApiSubscription
            responses:
                "200":
                    description: An API's subscription
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Subscription"
                default:
                    $ref: "#/components/responses/Error"
        put:
            tags:
                - API Subscriptions
            summary: Update one API's subscription
            description: |-
                Update the API's subscription.

                User must have the API_SUBSCRIPTION[UPDATE] permission.
            operationId: updateApiSubscription
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/UpdateSubscription"
                required: true
            responses:
                "200":
                    description: API's subscription successfully updated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Subscription"
                default:
                    $ref: "#/components/responses/Error"
...
    /environments/{envId}/apis/{apiId}/subscriptions/{subscriptionId}/_transfer:
        parameters:
            - $ref: "#/components/parameters/envIdParam"
            - $ref: "#/components/parameters/apiIdParam"
            - $ref: "#/components/parameters/subscriptionIdParam"
        post:
            tags:
                - API Subscriptions
            summary: Transfer an API subscription
            description: |-
                Transfer an API subscription to another plan.<br>
                The targeted plan must be published and must have the same security type. Transfer cannot be performed if general conditions apply to the targeted plan.

                User must have the API_SUBSCRIPTION[UPDATE] permission.
            operationId: transferApiSubscription
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/TransferSubscription"
                required: true
            responses:
                "200":
                    description: API's subscription successfully transferred
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/Subscription"
                default:
                    $ref: "#/components/responses/Error"
...
components:
    schemas:
...
        BaseApi:
            type: object
            properties:
                id:
                    type: string
                    description: API's uuid.
                    example: 00f8c9e7-78fc-4907-b8c9-e778fc790750
                name:
                    type: string
                    description: API's name. Duplicate names can exists.
                    example: My Api
                description:
                    type: string
                    description: API's description. A short description of your API.
                    example: I can use many characters to describe this API.
...
        BasePlan:
            type: object
            properties:
                id:
                    type: string
                    description: Plan's uuid.
                    example: 00f8c9e7-78fc-4907-b8c9-e778fc790750
                name:
                    type: string
                    description: Plan's name. Duplicate names can exists.
                    example: My Api plan
                description:
                    type: string
                    description: Plan's description. A short description of your plan.
                    example: I can use a hundred characters to describe this plan.
                apiId:
                    type: string
                    description: Id of the API owning the plan.
                    example: 6c530064-0b2c-4004-9300-640b2ce0047b
                security:
                    $ref: "#/components/schemas/PlanSecurity"
...
        BaseSubscription:
            type: object
            properties:
                id:
                    type: string
                    description: Subscription's uuid.
                    example: 34f8c9e7-68fd-4922-b8c9-e778fc790777
        Subscription:
            allOf:
                - $ref: "#/components/schemas/BaseSubscription"
                - properties:
                      api:
                          $ref: "#/components/schemas/BaseApi"
                      plan:
                          $ref: "#/components/schemas/BasePlan"
                      application:
                          $ref: "#/components/schemas/BaseApplication"
                      consumerMessage:
                          type: string
                          description: Message given by the api consumer when subscribing to the api.
                      publisherMessage:
                          type: string
                          description: Message given by the api publisher when accepting or rejecting the subscription.
                      metadata:
                          type: object
                          description: A list of metadata associated to this subscription.
                          additionalProperties:
                              type: string
                      daysToExpirationOnLastNotification:
                          type: integer
                          format: int32
                          description: Number of days before the expiration of this subscription when the last pre-expiration notification was sent.
                      consumerConfiguration:
                          $ref: "#/components/schemas/SubscriptionConsumerConfiguration"
                      failureCause:
                          type: string
                          description: Details about the last failure encountered on this subscription.
                      status:
                          $ref: "#/components/schemas/SubscriptionStatus"
                      consumerStatus:
                          $ref: "#/components/schemas/SubscriptionConsumerStatus"
                      processedBy:
                          $ref: "#/components/schemas/BaseUser"
                      subscribedBy:
                          $ref: "#/components/schemas/BaseUser"
                      processedAt:
                          type: string
                          format: date-time
                          description: The datetime when the subscription was processed.
                          example: 2023-05-18T12:40:46.184Z
                      startingAt:
                          type: string
                          format: date-time
                          description: The datetime when the subscription starts. No starting date means the subscription starts immediately.
                          example: 2023-05-20T12:30:42.101Z
                      endingAt:
                          type: string
                          format: date-time
                          description: The datetime when the subscription ends. No ending date means the subscription never ends.
                          example: 2023-05-25T12:40:46.184Z
                      createdAt:
                          type: string
                          format: date-time
                          description: The datetime when the subscription was created.
                          example: 2023-05-18T12:40:46.184Z
                      updatedAt:
                          type: string
                          format: date-time
                          description: The last datetime when the subscription was updated.
                          example: 2023-05-25T12:40:46.184Z
                      closedAt:
                          type: string
                          format: date-time
                          description: The datetime when the subscription was closed.
                          example: 2023-05-31T13:40:46.184Z
                      pausedAt:
                          type: string
                          format: date-time
                          description: The datetime when the subscription was paused by the api publisher.
                          example: 2023-05-25T12:40:46.184Z
                      consumerPausedAt:
                          type: string
                          format: date-time
                          description: The datetime when the subscription was paused by the api consumer.
                          example: 2023-05-25T12:40:46.184Z
        SubscriptionStatus:
            type: string
            description: The status of the subscription manageable by the api publisher.
            example: ACCEPTED
            enum:
                - PENDING
                - REJECTED
                - ACCEPTED
                - CLOSED
                - PAUSED
                - RESUMED
                - INTEGRATION_PROCESSING
...
        SubscriptionConsumerConfiguration:
            type: object
            description: Consumer configuration associated to the subscription in case it is attached to a push plan.
            properties:
                entrypointId:
                    type: string
                    description: The id of the targeted entrypoint
                    example: webhook
                channel:
                    type: string
                    description: The channel to consume
                    example: /channel1
                entrypointConfiguration:
                    type: object
                    description: The configuration to use at subscription time to push to the target service.
                    example: |
                        {
                            "callbackUrl": "https://webhook.site/aaaaaaaa-ffff-ffff-ffff-aaaaaaaa",
                            "headers": [
                                {
                                    "name": "demoHeader",
                                    "value": "my-value"
                                },
                                {
                                    "name": "anotherHeader",
                                    "value": "my-value2"
                                }
                            ]
                        }
            required: [entrypointId]
...
        UpdateSubscription:
            type: object
            properties:
                consumerConfiguration:
                    $ref: "#/components/schemas/SubscriptionConsumerConfiguration"
                metadata:
                    type: object
                    description: A list of metadata associated to this subscription.
                    additionalProperties:
                        type: string
                startingAt:
                    type: string
                    format: date-time
                    description: The datetime when the subscription starts. No starting date means the subscription starts immediately.
                    example: 2023-05-20T12:30:42.101Z
                endingAt:
                    type: string
                    format: date-time
                    description: The datetime when the subscription ends. No ending date means the subscription never ends.
                    example: 2023-05-25T12:40:46.184Z
...
        TransferSubscription:
            type: object
            properties:
                planId:
                    type: string
                    description: The plan identifier to transfer the subscription to.
...
        BaseApplication:
            type: object
            properties:
                id:
                    type: string
                    description: Application's uuid.
                    example: 00f8c9e7-78fc-4907-b8c9-e778fc790750
                name:
                    type: string
                    description: Application's name. Duplicate names can exist.
                    example: My App
                description:
                    type: string
                    description: Application's description. A short description of your App.
                    example: I can use a hundred characters to describe this App.
                domain:
                    type: string
                    description: Application's domain. Domain used by application, if relevant.
                    example: https://my-app.com
                type:
                    type: string
                    description: Application's type.
                    example: iOS
                primaryOwner:
                    $ref: "#/components/schemas/PrimaryOwner"
                apiKeyMode:
                    $ref: "#/components/schemas/ApiKeyMode"
...
    parameters:
...
        envIdParam:
            name: envId
            in: path
            required: true
            description: Id or Hrid (Human readable Id) of an environment.
            schema:
                type: string
                default: DEFAULT
        apiIdParam:
            name: apiId
            in: path
            required: true
            description: Id of an API.
            schema:
                type: string
...
        subscriptionIdParam:
            name: subscriptionId
            in: path
            required: true
            description: Id of a subscription.
            schema:
                type: string
...
    responses:
...
        Error:
            description: Generic error response
            content:
                application/json:
                    schema:
                        $ref: "#/components/schemas/Error"
...
    securitySchemes:
        BasicAuth:
            type: http
            scheme: basic
        CookieAuth:
            type: apiKey
            in: cookie
            name: Auth-Graviteeio-APIM